Refactor feature_extractor.py to encoder module
Renamed FeatureExtractor to Encoder
